Fix AEP layer timing

+++ src/core/io/aep/riff.hpp
@@ -269,7 +269,7 @@ public:
     std::uint16_t read_uint16() { return read_uint<2>(); }
     std::uint32_t read_uint32() { return read_uint<4>(); }
 
-    std::uint32_t read_sint16() { return read_sint<2>(); }
+    std::int16_t read_sint16() { return read_sint<2>(); }
     std::uint32_t read_sint32() { return read_sint<4>(); }
 
     float read_float32()
